backend/requirements/base.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-auditlog==3.0.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-decouple==3.8
Pillow==10.4.0
openpyxl==3.1.5
reportlab==4.2.5
